Doc: describe how to build/deploy OpenCL applications
Please see doc-funky/build-opencl-apps.md.

+++ funky-scripts/examples/build.sh
@@ -1,14 +1,42 @@
 #!/bin/bash
 BUILD_DIR=$1
 
+function usage {
+  cat <<EOF
+Usage: 
+  $(basename ${0}) <build dir>
+
+Options: 
+  -h                    print help
+EOF
+}
+
+########### get arguments #############
+while getopts h OPT
+do
+  case $OPT in
+    "h" )
+      usage
+      exit -1 ;;
+  esac
+done
+
 if [ -z ${BUILD_DIR} ]; then
-  echo 'Usage: ./build.sh <build_dir>'
+  usage
   exit -1
 fi
 
 if [ -e ${BUILD_DIR} ]; then
-  echo "Warning: the selected directory ${BUILD_DIR}/ already exists. This script removes the old one and re-compile it."
-  rm -r ${BUILD_DIR}
+  echo "Warning: the selected directory ${BUILD_DIR}/ already exists. This script will remove the old one and re-compile it. "
+  read -p "Do you continue?: [y/N]: " yn
+  case ${yn} in
+    [yY]*) 
+      rm -r ${BUILD_DIR}
+      ;;
+    *) 
+      echo "Abort."
+      exit -1 ;;
+  esac
 fi
 
 mkdir ${BUILD_DIR}
